Moved NameAndTitle from GUI to Engine layer
NameAndTitle functions had no GUI dependencies - they only used the
Engine-level localization global. Moved declarations to Localization.h
and implementations to Localization.cpp, and simplified
Actor::GetDisplayName() to use NameAndTitle(getNPCData(npcId)).
